+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ## [7.5.2] - 2024-10-23
 ### Fixed
 - Added condition for `k8s` service account secret creation.
+- Added new variable `ecs_platform_version` to handle ECS platform version update.
 
 ## [7.5.1] - 2024-10-22
 ### Added